+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/Formats.java
@@ -1,7 +1,8 @@
 package net.splitcells.website;
 
 public enum Formats {
-    HTML("text/html");
+    HTML("text/html")
+    , CSS("text/css");
 
     private final String mimeTypes;